+++ BrickBreaker/screens/gameScreen.hpp
@@ -54,7 +54,7 @@ namespace BrickBreaker
 		void Start() override;
 		void AddNewEntities() override;
 		void RemoveOldEntities() override;
-		std::uint16_t Update(std::chrono::microseconds elapsedTime) override;
+		std::uint64_t Update(std::chrono::microseconds elapsedTime) override;
 		void Draw(std::chrono::microseconds elapsedTime) override;
 		void ProcessInput(std::chrono::microseconds elapsedTime) override;
 		void AddEntity(entities::EntityPtr entity);
